Add support for "realigning to first line"
Here, "realigning to first line" means that only the first line and lines whose
first token is of the same kind as the first token of the first line will be
realigned - other lines will be left unchanged. This makes it simple to align
correctly while including comment lines etc.